requirements.txt
requests>=2.28
pandas>=1.5
python-dotenv>=1.0
APScheduler>=3.10
